MigrationManager factored out.
